+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2 - 19 March 2025
+
+* Update version
+* bugfix scrolling issue
+* Add copy http record to clipboard
+
 ## 1.0.1 - 18 March 2025
 
 * Update version